RSS current MOS masks (#42)
MOS masks currently in the magazine
