+++ saltapi/main.py
@@ -13,10 +13,10 @@ from saltapi.settings import Settings
 from saltapi.web.api.authentication import router as authentication_router
 from saltapi.web.api.block_visits import router as block_visits_router
 from saltapi.web.api.blocks import router as blocks_router
-from saltapi.web.api.mos import router as mos_router
 from saltapi.web.api.proposals import router as proposals_router
 from saltapi.web.api.user import router as user_router
 from saltapi.web.api.users import router as users_router
+from saltapi.web.api.instruments import router as instruments_router
 
 app = FastAPI()
 
@@ -58,4 +58,4 @@ app.include_router(authentication_router)
 app.include_router(block_visits_router)
 app.include_router(user_router)
 app.include_router(users_router)
-app.include_router(mos_router)
+app.include_router(instruments_router)
