+++ saltapi/repository/instrument_repository.py
@@ -1,3 +1,5 @@
+from typing import List, Dict, Any
+
 from sqlalchemy.engine import Connection
 
 from saltapi.repository.bvit_repository import BvitRepository
@@ -29,3 +31,11 @@ class InstrumentRepository:
     def get_bvit(self, bvit_id: int) -> BVIT:
         """Return a BVIT setup."""
         return self.bvit_repository.get(bvit_id)
+
+    def get_mos_mask_in_magazine(self) -> List[str]:
+        """The list of MOS masks in the magazine."""
+        return self.rss_repository.get_mos_mask_in_magazine()
+
+    def get_mos_blocks(self, semesters: List[str]) -> List[Dict[str, Any]]:
+        """The list of MOS blocks."""
+        return self.rss_repository.get_mos_blocks(semesters)
