+++ saltapi/web/services.py
@@ -2,14 +2,13 @@ from sqlalchemy.engine import Connection
 
 from saltapi.repository.block_repository import BlockRepository
 from saltapi.repository.instrument_repository import InstrumentRepository
-from saltapi.repository.mos_repository import MosRepository
 from saltapi.repository.proposal_repository import ProposalRepository
 from saltapi.repository.target_repository import TargetRepository
 from saltapi.repository.user_repository import UserRepository
 from saltapi.service.authentication_service import AuthenticationService
 from saltapi.service.block_service import BlockService
+from saltapi.service.instrument_service import InstrumentService
 from saltapi.service.mail_service import MailService
-from saltapi.service.mos_service import MosService
 from saltapi.service.permission_service import PermissionService
 from saltapi.service.proposal_service import ProposalService
 from saltapi.service.user_service import UserService
@@ -60,7 +59,8 @@ def user_service(connection: Connection) -> UserService:
     return UserService(user_repository)
 
 
-def mos_service(connection: Connection) -> MosService:
-    """Return a MOS service instance."""
-    mos_repository = MosRepository(connection)
-    return MosService(mos_repository)
+def instrument_service(connection: Connection) -> InstrumentService:
+    """Return an instrument service instance."""
+    instrument_repository = InstrumentRepository(connection)
+    return InstrumentService(instrument_repository)
+
